cleanup, sync

+++ setup.py
@@ -5,6 +5,7 @@ __email__     = 'info@radical-cybertools.org'
 __copyright__ = 'Copyright 2013-20, The RADICAL-Cybertools Team'
 __license__   = 'MIT'
 
+
 ''' Setup script, only usable via pip. '''
 
 import re
@@ -15,13 +16,15 @@ import shutil
 
 import subprocess as sp
 
-from setuptools import setup, find_namespace_packages
+from setuptools import setup, Command, find_namespace_packages
 
 
 # ------------------------------------------------------------------------------
 name     = 'radical.utils'
 mod_root = 'src/radical/utils/'
 
+sdist_level = int(os.environ.get('SDIST_LEVEL', 0))
+os.environ['SDIST_LEVEL'] = str(sdist_level + 1)
 
 # ------------------------------------------------------------------------------
 #
@@ -123,7 +126,7 @@ def get_version(_mod_root):
             # the formerly derived version as ./VERSION
             shutil.move("VERSION", "VERSION.bak")              # backup
             shutil.copy("%s/VERSION" % _path, "VERSION")       # version to use
-            os.system  ("python3 setup.py sdist")               # build sdist
+            os.system  ("python3 setup.py sdist")              # build sdist
             shutil.copy('dist/%s' % _sdist_name,
                         '%s/%s'   % (_mod_root, _sdist_name))  # copy into tree
             shutil.move('VERSION.bak', 'VERSION')              # restore version
@@ -143,19 +146,29 @@ version, version_detail, sdist_name, path = get_version(mod_root)
 
 
 # ------------------------------------------------------------------------------
-#
-def read(*rnames):
+# check python version, should be >= 3.6
+if sys.hexversion < 0x03060000:
+    raise RuntimeError('ERROR: %s requires Python 3.6 or newer' % name)
 
+
+# ------------------------------------------------------------------------------
+#
+def read(fname):
     try:
-        return open(os.path.join(os.path.dirname(__file__), *rnames)).read()
+        return open(fname, encoding='utf-8').read()
     except Exception:
         return ''
 
 
 # ------------------------------------------------------------------------------
-# check python version, should be >= 3.6
-if sys.hexversion < 0x03060000:
-    raise RuntimeError('ERROR: %s requires Python 3.6 or newer' % name)
+#
+class RunTwine(Command):
+    user_options = []
+    def initialize_options(self): pass
+    def finalize_options(self):   pass
+    def run(self):
+        _, _, ret = sh_callout('python3 setup.py sdist upload -r pypi')
+        raise SystemExit(ret)
 
 
 # ------------------------------------------------------------------------------
@@ -233,7 +246,7 @@ setup_args = {
     'test_suite'         : '%s.tests' % name,
     'zip_safe'           : False,
     'data_files'         : df,
-    'cmdclass'           : {},
+    'cmdclass'           : {'upload': RunTwine},
 }
 
 
@@ -241,12 +254,14 @@ setup_args = {
 #
 setup(**setup_args)
 
+
 # ------------------------------------------------------------------------------
 # clean temporary files from source tree
-os.system('rm -vrf src/%s.egg-info' % name)
-os.system('rm -vf  %s/%s'           % (path, sdist_name))
-os.system('rm -vf  %s/VERSION'      % path)
-os.system('rm -vf  %s/SDIST'        % path)
+if sdist_level == 0:
+    os.system('rm -vrf src/%s.egg-info' % name)
+    os.system('rm -vf  %s/%s'           % (path, sdist_name))
+    os.system('rm -vf  %s/VERSION'      % path)
+    os.system('rm -vf  %s/SDIST'        % path)
 
 
 # ------------------------------------------------------------------------------
